+++ DemoApp/StreamChat/Components/DemoChatChannelListRouter.swift
@@ -532,6 +532,20 @@ final class DemoChatChannelListRouter: ChatChannelListRouter {
                 } catch {
                     self.rootViewController.presentAlert(title: error.localizedDescription)
                 }
+            }),
+            .init(title: "Reset User Image", handler: { [unowned self] _ in
+                do {
+                    let connectedUser = try self.rootViewController.controller.client.makeConnectedUser()
+                    Task {
+                        do {
+                            try await connectedUser.update(unset: ["image"])
+                        } catch {
+                            self.rootViewController.presentAlert(title: error.localizedDescription)
+                        }
+                    }
+                } catch {
+                    self.rootViewController.presentAlert(title: error.localizedDescription)
+                }
             })
         ])
     }